+++ src/main/java/com/achobeta/domain/users/controller/UserController.java
@@ -1,15 +1,13 @@
 package com.achobeta.domain.users.controller;
 
 import com.achobeta.common.SystemJsonResponse;
-import com.achobeta.domain.email.component.EmailValidator;
-import com.achobeta.domain.users.service.EmailService;
 import com.achobeta.domain.users.service.UserService;
-import com.achobeta.domain.users.util.IdentifyingCodeValidator;
-import com.achobeta.exception.IllegalEmailException;
-import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 /**
  * @author BanTanger 半糖
@@ -21,34 +19,7 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api/v1/user")
 public class UserController {
 
-    private final UserService userService;
-
-    private final EmailService emailService;
-
-    @PostMapping("/check")
-    public SystemJsonResponse emailIdentityCheck(@RequestParam("email") @NonNull String email) {
-        if(!EmailValidator.isEmailAccessible(email)) {
-            throw new IllegalEmailException("邮箱格式错误");
-        }
-        // 获得随机数
-        String code = IdentifyingCodeValidator.getIdentifyingCode();
-        emailService.sendIdentifyingCode(email, code);
-        // 能到这一步就成功了
-        log.info("发送验证码:{} -> email:{}", code, email);
-        return SystemJsonResponse.SYSTEM_SUCCESS();
-    }
-
-    @PostMapping("/check/{code}")
-    public SystemJsonResponse checkCode(@RequestParam("email") @NonNull String email,
-                                        @PathVariable("code") @NonNull String code) {
-        if(!EmailValidator.isEmailAccessible(email)) {
-            throw new IllegalEmailException("邮箱格式错误");
-        }
-        emailService.checkIdentifyingCode(email, code);
-        // 成功
-        log.info("email:{}, 验证码:{} 验证成功", email, code);
-        return SystemJsonResponse.SYSTEM_SUCCESS();
-    }
+    private UserService userService;
 
     /**
      * http://localhost:9001/login?name=bantanger
